+++ parser/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add built-in type `result`.
+
 ## [0.3.0] - 2024-01-18
 
 ### Added
@@ -22,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
+[Unreleased]: https://github.com/dennis-hamester/aldrin/compare/aldrin-parser-0.3.0...HEAD
 [0.3.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-parser-0.3.0
 [0.2.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-parser-0.2.0
 [0.1.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-parser-0.1.0